Generated the PCB to add fingerprint sensor to the keyboard
In order to add this new feature, a microcontroller was added.
It is duplicated because this board can be used as left and
right sides of the keyboard, so only one of the sides should
be soldered.

The Atmega microcontroller reads the fingerprint sensor using
UART protocol while the nRFmicro reads the Arduino fingerprint
sensor read data using I2C sensor. An additional wire was added
between the Atmega and the nRFmicro to the Atmega be able to
interrupt the nRFmicro every time the fingerprint sensor detects a
finger.